fixes (https://github.com/Macoron/whisper.unity/pull/25)
