--- Assets/Samples/5 - Initial Prompt.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9c9b720060358bf428634acbadf70799
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
