--- Assets/Samples/5 - Initial Prompt/5 - Initial Prompt.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 1116922a700363245b3973ca3c0e158f
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
