--- Assets/Samples/5 - Initial Prompt/InitialPromptDemo.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: b84fd2ad6f7348928f1a38d3c6cc59e7
-timeCreated: 1685308713
